requirements.txt
fastapi==0.115.4
gensim==4.3.3
grpcio==1.67.1
kafka==1.3.5
nltk==3.9.1
pydantic==2.9.2
pymongo==4.10.1
python-dotenv>=1.0.0
Requests==2.32.3
scikit_learn==1.5.2
spacy==3.7.2
textblob==0.18.0.post0
vaderSentiment==3.3.2
tqdm>=4.62.3
